fix: split oversized view-builder markers test for line-size gate (#3846)
The consolidated markers test exceeded the 1000 non-comment line cap;
extract civilian-marker policy and port/town co-location coverage into
a fourth view-builder test file (at the hard cap of 4).
